+++ lib/hexcolor.dart
@@ -1,1 +1,34 @@
-export 'src/src.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+
+class HexColor extends Color {
+  ///use hex color strings in your app
+  static const MethodChannel _channel = const MethodChannel('hexcolor');
+
+  static Future<String> get platformVersion async {
+    final String version = await _channel.invokeMethod('getPlatformVersion');
+    return version;
+  }
+
+  static int _getColorFromHex(String hexColor) {
+    hexColor = hexColor.toUpperCase().replaceAll("#", "");
+    if (hexColor.length == 6) {
+      hexColor = "FF" + hexColor;
+    }
+    return int.parse(hexColor, radix: 16);
+  }
+
+  HexColor(final String hexColor) : super(_getColorFromHex(hexColor));
+}
+
+class ColorToHex extends Color {
+  ///convert material colors to hexcolor
+  static int _convertColorTHex(Color color) {
+    var hex = '${color.value}';
+    return int.parse(
+      hex,
+    );
+  }
+
+  ColorToHex(final Color color) : super(_convertColorTHex(color));
+}